refactor: dto 패키지 생성
