+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,11 +1,12 @@
 package racingcar.controller;
 
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import racingcar.domain.RacingCars;
+import racingcar.dto.GameRequest;
+import racingcar.dto.GameResponse;
+import racingcar.dto.RacingCarDto;
 import racingcar.exception.CommaNotFoundException;
 import racingcar.service.RacingGameService;
