+++ src/main/java/racingcar/dao/JdbcRacingGameDao.java
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import racingcar.dto.PlayerSaveDto;
 import racingcar.exception.DatabaseAccessException;
 
 import java.util.List;
