+++ src/main/java/racingcar/dto/GameRequest.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
