+++ src/main/java/racingcar/dto/GameResponse.java
@@ -1,4 +1,4 @@
-package racingcar.controller;
+package racingcar.dto;
 
 import java.util.List;
