+++ src/main/java/racingcar/dto/PlayerSaveDto.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.dto;
 
 public class PlayerSaveDto {
