+++ src/test/java/racingcar/dao/JdbcRacingGameDaoTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import racingcar.dto.PlayerSaveDto;
 
 import java.util.List;
